requirements.txt
Flask
Flask-WTF
Flask-Login
python-dotenv
Flask-SQLAlchemy
Pydantic